+++ src/Smartstore.Core/Catalog/Discounts/Domain/DiscountEnums.cs
@@ -56,4 +56,34 @@
         /// </summary>
         NTimesPerCustomer = 25
     }
+
+    /// <summary>
+    /// Allows to specify which discount requirements to be validated.
+    /// </summary>
+    [Flags]
+    public enum DiscountValidationFlags
+    {
+        None = 0,
+
+        /// <summary>
+        /// Validates <see cref="Discount.DiscountLimitation"/>.
+        /// </summary>
+        WithDiscountLimitations = 1 << 0,
+
+        /// <summary>
+        /// Checks the shopping cart for the existence of gift cards.
+        /// No discount is applied if the cart contains gift cards because the customer could earn money through that.
+        /// </summary>
+        WithGiftCards = 1 << 1,
+
+        /// <summary>
+        /// Validates cart rules.
+        /// </summary>
+        WithRules = 1 << 2,
+
+        /// <summary>
+        /// Validates all discount requirements.
+        /// </summary>
+        All = WithDiscountLimitations | WithGiftCards | WithRules
+    }
 }