+++ src/Smartstore.Core/Platform/DataExchange/Export/DataExporter.cs
@@ -1554,7 +1554,7 @@ namespace Smartstore.Core.DataExchange.Export
             var priceDisplay = ctx.Projection.PriceType ?? _priceSettings.PriceDisplayType;
             var options = _priceCalculationService.CreateDefaultOptions(false, ctx.ContextCustomer, ctx.ContextCurrency, batchContext);
 
-            options.CheckDiscountValidity = true;
+            options.DiscountValidationFlags = DiscountValidationFlags.All;
             options.DetermineLowestPrice = priceDisplay == PriceDisplayType.LowestPrice;
             options.DeterminePreselectedPrice = priceDisplay == PriceDisplayType.PreSelectedPrice;
             options.ApplyPreselectedAttributes = priceDisplay == PriceDisplayType.PreSelectedPrice;